requirements.txt
# Raspberry Pi hardware control
gpiozero==2.0.1
adafruit-circuitpython-dht==4.0.7
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2

# Audio processing
sounddevice==0.5.1
soundfile==0.13.1
numpy==2.1.3

# OpenAI Whisper for speech-to-text
openai-whisper==20240930

# Testing
pytest==8.3.5
